requirements.txt
# ============================================
# VoiceFlow AI SaaS - Dependencies
# ============================================

# Core ML & Voice AI
torch==2.1.0
torchaudio==2.1.0
transformers==4.35.0
openai-whisper==20231117
librosa==0.10.1
soundfile==0.12.1
pydub==0.25.1
scipy==1.11.3

# Indian Languages Support
ai4bharat-transliteration==1.0.0
indic-nlp-library==0.91

# API Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
websockets==12.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Database
sqlalchemy==2.0.23
alembic==1.12.1
asyncpg==0.29.0
redis==5.0.1

# Authentication & Security
passlib[bcrypt]==1.7.4
pyjwt==2.8.0
email-validator>=2.1.0
slowapi>=0.1.9

# Telephony
twilio==8.10.0

# HTTP
httpx==0.25.2
aiohttp==3.9.1

# Monitoring
prometheus-client==0.19.0
sentry-sdk[fastapi]==1.35.0

# Utilities
python-dotenv==1.0.0
jinja2==3.1.2
python-dateutil==2.8.2
tenacity==8.2.3

# AI/LLM Providers
anthropic>=0.7.0
openai>=1.3.0
groq>=0.3.0

# Voice Synthesis (TTS)
deepgram-sdk>=2.11.0
edge-tts>=6.1.0
sentencepiece>=0.1.99
accelerate>=0.25.0

# Payments (India)
razorpay>=1.4.0

# Security
cryptography>=41.0.0

# Database Production
psycopg2-binary>=2.9.0

# Background Tasks
apscheduler>=3.10

# Async File Handling
aiofiles>=23.2.0
